requirements.txt
PyQt6
pyqt6-tools
pandas
tk
dfa
pandastable
nltk
automata-lib
pyside6